done with all changes

+++ backend/routes/course.routes.js
@@ -7,11 +7,8 @@ const productRouter=express.Router();
 
 productRouter.get("/getcoursesformylearning",async(req,res)=>{
     const {userId}=req.query;
-    // console.log("zz",req.query);
-    // console.log("z",userId);
     try{
         let purchasedCourse = await PurchasedModel.findOne({ userId });
-        // console.log("zzz",purchasedCourse)
           const courses = await ProductModel.find({ _id: { $in: purchasedCourse.productId } });
           console.log("courses",courses);
           res.status(200).send(courses);
@@ -194,8 +191,14 @@ productRouter.delete("/deletefromcart",async(req,res)=>{
     const userId=req.query.userId;
     const productId=req.query.productId;
     try{
-        await AddToCartModel.findOneAndDelete({userId,productId});
-        res.status(200).send({"message":"removed from cart"});
+        if (Array.isArray(productId)) {
+            await AddToCartModel.deleteMany({ userId, productId: { $in: productId } });
+            res.status(200).send({ message: "Products removed from cart" });
+          }
+           else {
+             await AddToCartModel.findOneAndDelete({userId,productId});
+            res.status(200).send({ message: "Product removed from cart" });
+          }
     }
     catch(err){
         res.status(400).send({"message":"error deleting from cart"});
@@ -221,9 +224,6 @@ productRouter.post("/addtopurchased",async(req,res)=>{
         purchasedCourse.productId = [...new Set([...purchasedCourse.productId, ...productId])];
       }
       await purchasedCourse.save();
-  
-      // Fetch course details for the purchased productIds
-    //   const courses = await ProductModel.find({ _id: { $in: purchasedCourse.productId } });
       res.status(200).json({ message: 'Purchased courses stored successfully.' });
     } 
     catch (error) {
